populate_base/elasticsearch/requirements.txt
elasticsearch
requests
python-binance